requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.4
django-extensions==3.1.5
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
executing==0.8.3
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.3
ipython==8.3.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
oauthlib==3.2.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stack-data==0.2.0
testfixtures==6.18.5
traitlets==5.2.1.post0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
